Extract @ExceptionHandler method into @ControllerAdvice class in order to allow easy overriding
